+++ changelog.md
@@ -156,6 +156,7 @@ Codion Change Log
 - EntityConditionPanelFactory now sets the string provider for the in values list box.
 - EntityTablePanel now removes the condition and filter panels when they are hidden, related refactoring.
 - EntityTablePanel.createConditionPanel() replaced with Config.tableConditionPanelFactory(), conditionPanel no longer initialized lazily.
+- EntityTablePanel, condition panel field configuration moved to EntityFieldFactory.
 ### is.codion.swing.framework.ui.tools
 - DomainGeneratorPanel.DEFAULT_USERNAME configuration value added.
 - DomainGeneratorPanel.DEFAULT_USERNAME renamed DEFAULT_USER, can include password.